# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticación",
        "description": "API para la autenticación de usuarios",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/fermentations": {
            "get": {
                "tags": [
                    "Fermentations"
                ],
                "summary": "Obtener todas las fermentaciones",
                "operationId": "9eff9ff1b1a3a0c7bf42a226c3fa399d",
                "responses": {
                    "200": {
                        "description": "Lista de fermentaciones"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Fermentations"
                ],
                "summary": "Crear una nueva fermentación",
                "operationId": "930ae03cf1c6c474f0fbc782bfd48e5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "device_id",
                                    "start_time",
                                    "status",
                                    "type",
                                    "genotypes"
                                ],
                                "properties": {
                                    "device_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-06-14T10:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "Especial",
                                            "Premium"
                                        ],
                                        "example": "Premium"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Nota de prueba",
                                        "nullable": true
                                    },
                                    "genotypes": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "number",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fermentación creada exitosamente"
                    },
                    "400": {
                        "description": "Error de validación o lógica de negocio"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/fermentations/{id}": {
            "get": {
                "tags": [
                    "Fermentations"
                ],
                "summary": "Obtener una fermentación por ID",
                "operationId": "5fd1a45bcc2330dae67d929f7afd29b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la fermentación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fermentación encontrada"
                    },
                    "404": {
                        "description": "Fermentación no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/fermentations/{id}/measurements": {
            "get": {
                "tags": [
                    "Fermentations"
                ],
                "summary": "Obtener las mediciones de una fermentación",
                "operationId": "c92d57258f42a6d542aefff6c8c85c74",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la fermentación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mediciones encontradas"
                    },
                    "404": {
                        "description": "Fermentación no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/fermentations/{id}/measurements/{limit}": {
            "get": {
                "tags": [
                    "Fermentations"
                ],
                "summary": "Obtener las últimas mediciones de una fermentación con límite",
                "operationId": "3692ec3448ef58f434640e6d91cdd702",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la fermentación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Cantidad máxima de registros a devolver",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mediciones encontradas"
                    },
                    "404": {
                        "description": "Fermentación no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/fermentations/{id}/status": {
            "put": {
                "tags": [
                    "Fermentations"
                ],
                "summary": "Actualizar el estado de una fermentación",
                "operationId": "602174340f0c846528054bd04dd72d7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la fermentación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado"
                    },
                    "404": {
                        "description": "Fermentación no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}